Add MemberData and MemberLevelData classes
New classes are added to implement a data model for storing member data. The MemberData class includes fields like id, username and created_at. In addition to these, until a member's level details such as xp, level, and the time stamp of the last update are stored in the MemberLevelData class. This addition should increase the capability to manipulate and manage member details efficiently.