Refine user data handling in ApiEventHandler
Updated the handling of user data in ApiEventHandler class to only execute the user level tracking functions if the "leveling" option is set to true in the guildConfig. This efficiency improvement was necessary to avoid unnecessary database queries and computation tasks when a bot is deployed across several data sets, some of them might not need leveling tracking.